luci-app-openvpn-server: tidy up luci (#8217)
