+++ package/lean/luci-app-openvpn-server/Makefile
@@ -5,17 +5,14 @@
 
 include $(TOPDIR)/rules.mk
 
+PKG_NAME:=luci-app-openvpn-server
+PKG_VERSION:=2.0
+PKG_RELEASE:=19
+
 LUCI_TITLE:=LuCI support for OpenVPN Server
 LUCI_DEPENDS:=+openvpn-openssl +openvpn-easy-rsa +kmod-tun
 LUCI_PKGARCH:=all
-PKG_NAME:=luci-app-openvpn-server
-PKG_VERSION:=2.0
-PKG_RELEASE:=18
 
 include $(TOPDIR)/feeds/luci/luci.mk
 
 # call BuildPackage - OpenWrt buildroot signature
-
-
-
-
