+++ package/lean/luci-app-openvpn-server/luasrc/view/openvpn-server/openvpn-server_status.htm
@@ -1,7 +1,7 @@
 <script type="text/javascript">//<![CDATA[
 XHR.poll(3, '<%=url([[admin]], [[vpn]], [[openvpn-server]], [[status]])%>', null,
 	function(x, data) {
-		var tb = document.getElementById('openvpn_status');
+		var tb = document.getElementById('openvpn-server_status');
 		if (data && tb) {
 			if (data.running) {
 				var links = '<em><b><font color=green>OpenVPN Server <%:RUNNING%></font></b></em>';
@@ -16,7 +16,7 @@ XHR.poll(3, '<%=url([[admin]], [[vpn]], [[openvpn-server]], [[status]])%>', null
 </script>
 <style>.mar-10 {margin-left: 50px; margin-right: 10px;}</style>
 <fieldset class="cbi-section">
-	<p id="openvpn_status">
+	<p id="openvpn-server_status">
 		<em><%:Collecting data...%></em>
 	</p>
 </fieldset>
