+++ package/lean/luci-app-openvpn-server/po/zh-cn/openvpn-server.po
@@ -16,20 +16,17 @@ msgstr "端口"
 msgid "WAN DDNS or IP"
 msgstr "WAN口的 DDNS域名 或者 IP"
 
-
 msgid "Client Network"
 msgstr "客户端网段"
 
 msgid "VPN Client Network IP with subnet"
 msgstr "客户端分配的网段地址（默认为 10.8.0.0 255.255.255.0）"
 
-
 msgid "Client Settings"
 msgstr "客户端推送配置"
 
 msgid "OpenVPN Client config file"
 msgstr "OpenVPN 客户端配置文件"
-
 
 msgid "Download .ovpn file"
 msgstr "一键下载 .ovpn 文件"
